Social: Move PostPreview to its only consumer directory
Move post-preview.tsx from the now-empty social-post-modal directory
to customize-and-preview/preview-section where it is imported.
